// src/app/models/request.ts
export interface Internal{
  projecttitle :string;
  projectname:string;
  projectdescription:string;
  teamnumber:string;
   email1 : string;
  email2: string;
  email3 : string;
  email4 : string;
  group1: string;
  group2: string;
  group3: string;
  group4: string;
  cms1: string;
  cms2: string;
  cms3: string;
  cms4: string;
  supName: string;
  cosupname:string


}